Correct the output quoting to match json standards
JSON requires double quotes around strings, but when printing lists and dictionaries, Python uses single quotes. The best solution would be to rewrite the sections which write the json output to use the json module, but as a stop-gap, just edit the output to replace single with double quotes where needed.